recipe in db should be working inshallah

+++ DietScoopApp/app/src/main/java/com/example/dietscoop/IngredientInRecipe.java
@@ -2,7 +2,7 @@ package com.example.dietscoop;
 
 public class IngredientInRecipe extends Ingredient{
 
-    public IngredientInRecipe(String description, String measurementUnit, int amount, Category category) {
+    public IngredientInRecipe(String description, String measurementUnit, double amount, Category category) {
         super(description,measurementUnit,amount, category);
         
     }
